refactor: 유저 정보 조회 쿼리 반환값 수정

+++ backend/src/main/java/com/ssafy/ssafyro/domain/report/ReportQueryRepository.java
@@ -8,6 +8,6 @@ public interface ReportQueryRepository {
 
     List<Report> findReports();
 
-    long countReportsType(RoomType type, User user);
+    int countReportsType(RoomType type, User user);
 
 }
